backend/requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.10.2
drf-yasg==1.16.1
idna==2.8
inflection==0.3.1
jsonschema==3.0.2
MarkupSafe==1.1.1
packaging==19.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
serializers==0.2.4
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
djangorestframework_simplejwt==4.3.0